move cloud yaml path test to unit tests
